client: Add first version of robot behavior
Implement first version of robot behavior. Currently, the robot listen for new
participants and, when a participant connect, start speech recognition.

The results of speech recognition are displayed in the console.

The robot requires the `start()` method to be called to begin. This method is
automatically called by the runner after loading all the modules.